shared module restructure
