+++ libs/angular/src/shared/components/password-strength/password-strength.component.ts
