+++ libs/angular/src/shared/components/password-strength/password-strength.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from "@angular/core";
+
+import { SharedModule } from "../../shared.module";
+
+import { PasswordStrengthComponent } from "./password-strength.component";
+
+@NgModule({
+  imports: [SharedModule],
+  declarations: [PasswordStrengthComponent],
+  exports: [PasswordStrengthComponent],
+})
+export class PasswordStrengthModule {}
